requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.1
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
